requirements.txt
ckanext-spatial @ git+https://github.com/ckan/ckanext-spatial.git#egg=ckanext-spatial
